feat: Adding the possible helped user list on help Offers
